requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.1.1
email-validator==2.2.0
python-dotenv==1.0.1
WTForms==3.1.2
itsdangerous==2.2.0
click==8.1.7
Flask-Talisman==1.1.0
